make two_pass the main impimentation
